[feature] Video player real-time collaboration implemented
