+++ src/components/VideoPlayer/VideoPlayer.tsx
@@ -50,7 +50,7 @@ type Props = {
 
 const VideoPlayer = ({ id, cam }: Props) => {
   const [app, setApp] = useState<TldrawApp>();
-
+  const [roomId, setRoomId] = useState<string>('');
   const [videoElemLoading, setVideoElemLoading] = useState<boolean>(true);
   const [videoPlayerLoading, setVideoPlayerLoading] = useState<boolean>(true);
   const [video, setVideo] = useState<Video>({} as Video);
@@ -77,6 +77,8 @@ const VideoPlayer = ({ id, cam }: Props) => {
 
     const el = document.createElement('video');
     el.src = videoData.recordUrl;
+
+    setRoomId(id + 'default');
 
     setVideo({
       bookmarks: videoData.bookmarks
@@ -210,7 +212,8 @@ const VideoPlayer = ({ id, cam }: Props) => {
 
     const renderedDrawing = (
       <Drawing
-        onMount={(innerApp) => setApp(innerApp)}
+        app={app}
+        setApp={setApp}
         scale={1}
         playing={playing}
         video={video}
